remove the extra ".0" from the percentages in the temperature powers

+++ src/main/java/hydrologistmod/powers/ColdPower.java
@@ -42,7 +42,7 @@ public class ColdPower extends AbstractPower implements CloneablePowerInterface,
 
     @Override
     public void updateDescription() {
-        description = DESCRIPTIONS[0] + (100 - (calculateMultiplier() * 100)) + DESCRIPTIONS[1] + amount + (amount > 1 ? DESCRIPTIONS[3] : DESCRIPTIONS[2]) + DESCRIPTIONS[4] + (amount * 2) + DESCRIPTIONS[5];
+        description = DESCRIPTIONS[0] + (int)(100 - (calculateMultiplier() * 100)) + DESCRIPTIONS[1] + amount + (amount > 1 ? DESCRIPTIONS[3] : DESCRIPTIONS[2]) + DESCRIPTIONS[4] + (amount * 2) + DESCRIPTIONS[5];
     }
 
     @Override
